Assignment 5 - Navigation
Created a folder containing screens, changed the main directory of the app so it takes you to the "HomePage" has the regular To Do list, Form to add new tasks, but now has a button that will take you to the about page that tells you the name of the app and my name the creator. Also changed the color scheme to a more natural green and brown.